Update solution for gradient-descent-quadratic

+++ gradient-descent-quadratic/gradient-descent-quadratic.py
@@ -5,7 +5,7 @@ def gradient_descent_quadratic(a, b, c, x0, lr, steps):
     # Write code here
     import numpy as np 
 
-    x = x0
+    x = np.float64(x0)
 
     for i in range(steps):
       grad = 2*a*x + b
